Search and Add User Complete
