+++ models/search.js
@@ -12,12 +12,16 @@ import { conn } from "../connection.js"
   //Reviewed for post Team/Project Manager discussion split
   //Call to populate user card on search pagec when fully implemented
   async function getUserInfo(userEmail) {
-    let result = await conn.query(
-    "SELECT users.email as userEmail, users.first_name as firstName, users.last_name as lastName\
+    let [result] = await conn.query(
+    "SELECT users.id as userId, users.email as userEmail, users.first_name as firstName, users.last_name as lastName\
      FROM users WHERE users.email = ? limit 1",
     [userEmail]);
     return result;
   }
+
+
+
+
 
   async function checkIfUserExists(userId) {
     let [countData] = await conn.query(
@@ -31,24 +35,9 @@ import { conn } from "../connection.js"
     return false;
   }
   
-  async function addTeamMember(projectId, userId) {
-    let newTeamId = getTeamId(projectId)
-      try {
-        await conn.query(
-          "INSERT INTO project_users VALUES (?, ?) "
-          [newTeamId, userId]
-          );
-        return true;
-      } catch (error) {
-        return false;
-      }
-  }
+
   
-  async function getTeamId(projectId) {
-    let [result] = await conn.query("SELECT team_id FROM projects WHERE projectId = ? limit 1 ", [
-      projectId]);
-    return result;
-  }
+
 
   async function getTeamMembers(projectId) {
     let [result] = await conn.query("SELECT user_id FROM project_users WHERE projectId = ?", [
@@ -75,8 +64,6 @@ import { conn } from "../connection.js"
   export {
     checkIfUserExists,
     getUserInfo,
-    addTeamMember,
-    getTeamId,
     getTeamMembers,
     getUsersTeamId,
     getUsersTeams,
