+++ public/components/actionbar.ejs
@@ -3,7 +3,7 @@
     <a class="btn btn-dark <% if (page == 'project') {%> text-leaf-green <%}%>" href="/project/<%-project.id%>">Announcements</a>
     <a class="btn btn-dark <% if (page == 'tasks') {%> text-leaf-green <%}%>" href="/project/<%-project.id %>/tasks">Tasks</a>
     <a class="btn btn-dark <% if (page == 'schedule') {%> text-leaf-green <%}%>" href="/schedule">Schedule</a>
-    <a class="btn btn-dark <% if (page == 'team') {%> text-leaf-green <%}%>" href="/team/teams">Team</a>
+    <a class="btn btn-dark <% if (page == 'team') {%> text-leaf-green <%}%>" href="/search/<%-project.id %>">Add User</a>
     <!-- <button type="button" class="btn btn-dark" href="">+</button> -->
   </div>
 <% } %>
