+++ public/components/teams/search_card.ejs
@@ -1,0 +1,18 @@
+<div class="card" style="min-width: 300px; width: 300px;">
+  <div class="d-flex gap-1">
+    <%- user.firstName %>  <% user.lastName %>
+  </div>  
+  <img
+      class="card-img-top"
+      src="../../images/stock.jpg"
+      alt="Card image"
+      style="width: 100%"
+    />
+    <div class="card-body d-flex flex-column">
+      <a href="search/<%-projectId%>/searchUsers/add/<%-user.userId%>" class="btn btn-dark">Add User</a>
+        <a class="link-dark text-decoration-none mx-auto" style="width: fit-content;">
+        <i class="bi bi-lock-fill"></i>
+        <%- user.userEmail %>
+        </a>
+    </div>
+  </div>
