--- public/components/teams/team_card.ejs
@@ -1,15 +1,0 @@
-<div class="card" style="min-width: 300px; width: 300px;">
-    <img
-      class="card-img-top"
-      src="../../images/stock.jpg"
-      alt="Card image"
-      style="width: 100%"
-    />
-    <div class="card-body d-flex flex-column">
-      <a href="../../project/<%- projectId %>" class="btn btn-dark"><% if (locals.userEmail) {%><%- userEmail %><% } %></a><br />
-          <a href="../../team" class="link-dark text-decoration-none mx-auto" style="width: fit-content;">
-          <i class="bi bi-lock-fill"></i>
-          <%- email %>
-          </a>
-    </div>
-  </div>
